Fix error message for wrong output size
I bumped into this error and was very confused about how I could've
passed in anything other than a Uint8Array. After looking at the code
that was throwing the error, I think the error message is wrong.

This commit changes the error message so that it announces the size of
the array that it needs.